requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib
bcrypt
python-jose
python-multipart
pydantic-settings
python-dotenv
pandas